Log results when generating list of images on node
This is a troubleshooting commit only, so that I can figure out why our
CI node is unhappy. The VM where this will be executed is ephemeral and
part of the CI system, and has no sensitive data. So I'm not concerned
about the list of images being exposed for this particular
troubleshooting purpose.

Modifying this in Kubespray to assist in debugging
https://github.com/NVIDIA/deepops/pull/405